exit finally works under all situations

+++ src/ContactManager.java
@@ -11,6 +11,7 @@ import java.util.Scanner;
 public class ContactManager {
     public static void main(String[] args) {
 
+        boolean addedContact = false;
         System.out.println("1. View Contacts.\n" +
                 "2. Add a new contact.\n" +
                 "3. Search a contact by name.\n" +
@@ -18,15 +19,16 @@ public class ContactManager {
                 "5. Exit.\n" +
                 "Enter an option (1, 2, 3, 4 or 5):");
         Scanner scan = new Scanner(System.in);
-        String response = scan.next();
+        String response = scan.nextLine();
+        System.out.println(response);
 
-        if (response.contains("1")) {
+        if (response.equals("1")) {
             Contacts.main();
             System.out.println();
             System.out.println("Press any key to return to menu.");
             String back = scan.next();
                 main(null);
-        } else if (response.contains("2")) {
+        } else if (response.equals("2")) {
             String newContact;
             System.out.println("Enter first name of contact: ");
             String newName = scan.next();
@@ -34,19 +36,62 @@ public class ContactManager {
             String newNumber = scan.next();
             newContact = newName + " " + newNumber;
 
+            Path contactPath = Paths.get("src/", "Contacts.txt");
+            List<String> contactList = null;
             try {
-                Files.write(
-                        Paths.get("src/", "Contacts.txt"),
-                        Arrays.asList(newContact), // list with one item
-                        StandardOpenOption.APPEND
-                );
+                contactList = Files.readAllLines(contactPath);
             } catch (IOException e) {
                 e.printStackTrace();
             }
-            main(null);
+            for (int i = 0; i < contactList.size(); i++) {
+                String checkName = contactList.get(i);
+                String[] checkSplitName = checkName.split(" ");
+                if (checkSplitName[0].equalsIgnoreCase(newName)) {
+                    System.out.printf("There's already a contact named %s. Do you want to overwrite it? (Yes/No)", newName);
+                    String replaceConfirm = scan.next();
+                    if (replaceConfirm.equalsIgnoreCase("yes")) {
+
+                        try {
+                            contactList.remove(contactList.get(i));
+                            PrintWriter writer = new PrintWriter("src/contacts.txt");
+                            for (int ii = 0; ii < contactList.size(); ii++)
+                                writer.println(contactList.get(ii));
+                            writer.close();
+                            Files.write(
+                                    Paths.get("src/", "Contacts.txt"),
+                                    Arrays.asList(newContact), // list with one item
+                                    StandardOpenOption.APPEND
+                            );
+                        } catch (IOException e) {
+                            e.printStackTrace();
+                        }
+                        addedContact = true;
+                        System.out.println("Contact successfully replaced");
+                        main(null);
+                        break;
+                    } else if (replaceConfirm.equalsIgnoreCase("n")) {
+                        main(null);
+                        break;
+                    } else {
+                        System.out.println("Please enter a valid answer");
+                    }
+                }
+            }
+            if (!addedContact) {
+                try {
+                    Files.write(
+                            Paths.get("src/", "Contacts.txt"),
+                            Arrays.asList(newContact), // list with one item
+                            StandardOpenOption.APPEND
+                    );
+                } catch (IOException e) {
+                    e.printStackTrace();
+                }
+                main(null);
+            }
 
 
-        } else if (response.contains("3")) {
+        } else if (response.equals("3")) {
             do {
                 String newSearch;
                 System.out.println("Search for your contact.");
@@ -86,7 +131,7 @@ public class ContactManager {
                     }
                 } while (asking);
             } while (true);
-        } else if (response.contains("4")) {
+        } else if (response.equals("4")) {
             System.out.println("Who do you wish to shun?");
             String delete = scan.next();
             Path contactPath = Paths.get("src/", "Contacts.txt");
@@ -133,8 +178,9 @@ public class ContactManager {
                 main(null);
             }
 
-        } else if (response.contains("5")) {
+        } else if (response.equals("5")) {
             System.out.println("LATERSSSSSSSSSSS \uD83E\uDD2A");
+            System.exit(0);
         }
     }
 }
